don't include a main() function in base_local_planner library (#969)
